[ADDED] System channel to be notified of new channels
EXPERIMENTAL:

The server creates a system channel (_SYS.streaming.channels) that
applications can subscribe to like any other channel. On this
channel, the server will post an event when a new channel is created.
The even is JSON encoded and may look like this:

{"new":true,"name":"foo"}

Users can set limits on this channel as any other channel.