[EXPERIMENTAL] [ADDED] Encryption at rest for messages
Added a new store that adds encryption of messages payload capabilities
to existing store implementations.
Server will use encryption for any selected store type if started
with the Encryption option set to true and the EncryptionKey string.
If the key is not specified in the Options, the server will check
the environment variable NATS_STREAMING_ENCRYPTION_KEY and use this
if set.