[ADDED] Ability to delete channels through monitoring endpoints
A master switch enables access to the feature of deleting channels
from the monitoring endpoint, such as:
```
curl -X DELETE http://localhost:8222/channelsz/foo
```
However, the server will reject deletiion of a channel if there
is any registered subscription (durable or not, offline or not).
The user needs to ensure that no subscription is active, and
unsubscribe durable subsriptions.
Once the channel is deleted, it can be recreated with the same
name.